Screw it. Just use "double" instead of "long double".
I see no difference in the rendered images,
nor can I calculate a realistic scenario where it would make a difference,
but I DO see a MASSIVE difference in the time these programs
take.  Next to multi-threading, this is the biggest improvement.

...Later I might make the long-double be a compile-time option.
If you need it, you're probably looking at a part of the set that's
so small and complicated that you'd need a week to render it anyway.

+++ include/complex_helpers.h
@@ -40,7 +40,7 @@
 #include <stdbool.h>
 
 /* TODO: Some CLEAN way of making this a program-by-program option */
-typedef long double mfloat_t;
+typedef double mfloat_t;
 
 typedef struct complex_t {
         mfloat_t re;
